+++ components/organisms/TitleCenter.vue
@@ -51,3 +51,7 @@ export default {
   },
 }
 </script>
+
+<style lang="scss">
+@import '@/assets/styles/app';
+</style>
